Update random_code_decoupling_thm.yml
I made the above recommended changes and also made a subtle change to the description to express that the decoupling is approximate. I believe that all of the text is original, however, I presented the same intuitive explanation for decoupling as in Preskill's chapter 10 lecture notes and so I added a citation.