Update two_dimensional_hyperbolic_surface entry
Added a new bullet under 'notes:' on the existence of hyperbolic color codes which could simplify the implementation of logical gates. Let me know if I should move this to the fault-tolerant or general gates section. I also added hyperbolic color codes with reference to the QEC Zoo.

For your comment about the non-normalizability of oscillator code states, I'm not sure how to relate this to the hyperbolic code. Is there a way for these surface codes can be reconfigured for CV codes?